+++ heredoc.c
@@ -27,87 +27,65 @@ char	*hd_filename(t_data *shell, int count)
 	return (file);
 }
 
-static int	process_line(char *line, t_data *tokens, char *delimiter, int fd)
+static char	*extract_delimiter(char *line, int *i)
 {
-	if (ft_strcmp(line, delimiter) == 0)
+	int		start;
+	int		len;
+	char	*delimiter;
+
+	len = 0;
+	while (check_space(line[*i]))
+		(*i)++;
+	start = *i;
+	while (line[*i] && line[*i] != 31 && !check_space(line[*i]))
 	{
-		free(line);
-		return (1);
+		if (line[*i] == '\'' || line[*i] == '\"')
+			remove_quotes(&line[*i]);
+		len++;
+		(*i)++;
 	}
-	if (ft_strchr(line, '$'))
-		line = expand_var(tokens, line);
-	ft_putendl_fd(line, fd);
-	free(line);
-	return (0);
+	delimiter = (char *)malloc(sizeof(char) * (len + 1));
+	if (!delimiter)
+		return (NULL);
+	ft_strncpy(delimiter, &line[start], len);
+	delimiter[len] = '\0';
+	return (delimiter);
 }
 
-static int	handle_error(int stdin)
+static int	handle_heredoc(char *line, int *i, t_data *tokens)
 {
-	if (errno && errno == EBADF)
+	tokens->hd_count++;
+	*i += 2;
+	tokens->hd_delimeter = extract_delimiter(line, i);
+	if (!tokens->hd_delimeter)
 	{
-		dup2(stdin, STDIN_FILENO);
-		close(stdin);
+		error_message(tokens, "Memory allocation error\n", 1);
 		return (-1);
 	}
-	return (0);
-}
-
-static int	hd_readline(char *line, t_data *tokens, char *delimeter, int fd)
-{
-	int	stdin;
-
-	errno = 0;
-	stdin = dup(STDIN_FILENO);
-	if (stdin == -1)
-		return (-1);
-	while (1)
-	{		
-		if (g_signal_status)
-		{
-			free(line);
-			dup2(stdin, STDIN_FILENO);
-			close(stdin);
-			return (-1);
-		}
-		line = readline("> ");
-		if (!line)
-			break ;
-		if (process_line(line, tokens, delimeter, fd))
-			break ;
-	}
-	return (handle_error(stdin));
-}
-
-int	process_hd(t_data *tokens, const char *file, char *delimeter)
-{
-	int		fd;
-	char	*line;
-
-	line = NULL;
-	get_signal(HEREDOC);
-	unlink(file);
-	fd = open(file, O_RDWR | O_CREAT | O_TRUNC, 0644);
-	if (fd == -1)
-		error_message(tokens, "Failed to open fd for heredoc", 1);
-	if (hd_readline(line, tokens, delimeter, fd) == -1)
+	tokens->tempfile_hd = hd_filename(tokens, tokens->hd_count);
+	if (!tokens->tempfile_hd)
 	{
-		close (fd);
-		unlink(file);
+		error_message(tokens, "Failed to assign filename for heredoc", 1);
+		free(tokens->hd_delimeter);
+		tokens->hd_delimeter = NULL;
 		return (-1);
 	}
-	if (close(fd) == -1)
-		error_message(tokens, "Failed to close fd for heredoc", 1);
-	get_signal(DEFAULT);
+	if (process_hd(tokens, tokens->tempfile_hd, tokens->hd_delimeter) == -1)
+	{
+		free(tokens->hd_delimeter);
+		tokens->hd_delimeter = NULL;
+		return (-1);
+	}
+	free(tokens->hd_delimeter);
+	tokens->hd_delimeter = NULL;
 	return (0);
 }
 
 void	*is_heredoc(char *line, t_data *tokens)
 {
 	int		i;
-	int		j;
-	int		len;
 	int		in_quote;
-	
+
 	i = 0;
 	in_quote = 0;
 	tokens->hd_count = 0;
@@ -118,43 +96,12 @@ void	*is_heredoc(char *line, t_data *tokens)
 		{
 			in_quote = !in_quote;
 			i++;
-			continue;
+			continue ;
 		}
 		if ((ft_strncmp(&line[i], "<<", 2) == 0) && !in_quote)
-		{	
-			tokens->hd_count++;
-			i = i + 2;
-			while (check_space(line[i]))
-				i++;
-			j = i;
-			len = 0;
-			while (line[j] && line[j] != 31 && !check_space(line[j]))
-			{
-				len++;
-				j++;
-			}
-			tokens->hd_delimeter = (char *)malloc(sizeof(char) * len + 1);
-			if (tokens->hd_delimeter == NULL)
-				error_message(tokens, "Memory allocation error\n", 1);
-			j = 0;
-			while (line[i] && line[i] != 31 && !check_space(line[i]))
-			{
-				if (line[i] == '\'' || line[i] == '\"')
-					remove_quotes(line);
-				tokens->hd_delimeter[j++] = line[i++];
-			}
-			tokens->hd_delimeter[j] = '\0';
-			tokens->tempfile_hd = hd_filename(tokens, tokens->hd_count);
-			if (!tokens->tempfile_hd)
-				error_message(tokens, "Failed to assign filename for heredoc", 1);
-			if (process_hd(tokens, tokens->tempfile_hd, tokens->hd_delimeter) == -1)
-			{
-				free(tokens->hd_delimeter);
-				tokens->hd_delimeter = NULL;
+		{
+			if (handle_heredoc(line, &i, tokens) == -1)
 				break ;
-			}
-			free(tokens->hd_delimeter);
-			tokens->hd_delimeter = NULL;
 		}
 		if (line[i] != '\0')
 			i++;
