+++ minishell.h
@@ -32,7 +32,6 @@ extern int	g_signal_status;
 
 ///////// input validation /////////
 
-
 int				in_redir(t_data *shell, char *input);
 int				in_pipes(t_data *shell, char *input);
 bool			empty_line(char *line);
@@ -47,7 +46,7 @@ void			remove_quotes(char *str);
 int				skip_quotes(char *line, int i);
 int				q_pipe(t_data *shell, char *line);
 int				is_in_quotes(char c, int *in_single_quote,
-					int *in_double_quote);
+						int *in_double_quote);
 void			error_message(t_data *shell, char *msg, int status);
 void			msg_status(t_data *shell, char *msg, int status);
 void			init_t_data(t_data *data);
@@ -59,6 +58,8 @@ char			**do_split(char const *s, char c);
 void			split_line(char *line, t_data *shell);
 void			*is_heredoc(char *line, t_data *tokens);
 char			*hd_filename(t_data *shell, int count);
+int				process_hd(t_data *tokens, const char *file, 
+						char *delimeter);
 char			*expand_var(t_data *shell, char *line);
 void			make_redirs(t_data *tokens);
 void			remove_redir_from_input(t_data *tokens);
